requirements.txt
flask==2.3.2
flask-Mail==0.9.1
flask-SQLAlchemy==3.0.5
matplotlib==3.7.1
numpy==1.23.5
pandas==1.5.3
polars==0.18.4
pycaret==3.0.2
joblib==1.2.0
requests==2.31.0
google-cloud-storage==2.10.0